+++ src/pxblat/__init__.py
@@ -6,6 +6,7 @@ from .parser import read
 from .server import check_port_in_use
 from .server import check_port_open
 from .server import Client
+from .server import ClientThread
 from .server import create_client_option
 from .server import create_server_option
 from .server import DEFAULT_PORT
@@ -23,9 +24,14 @@ from .server import stop_server
 from .server import wait_server_ready
 from .toolkit import fa_to_two_bit
 from .toolkit import two_bit_to_fa
+from rich.traceback import install
+
+install(show_locals=True)
+
 
 __all__ = [
     "Server",
+    "ClientThread",
     "Client",
     "files",
     "server_query",