+++ src/pxblat/server/utils.py
@@ -1,7 +1,6 @@
 import logging
 import multiprocessing
 
-
 logger = multiprocessing.log_to_stderr()
 
 logger.setLevel(logging.WARNING)