+++ tests/test_result.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 import subprocess
 import sys
 import time
@@ -6,7 +7,10 @@ from pathlib import Path
 
 import pxblat
 import pytest
+from pxblat import Client
 from rich import print
+
+from Bio import SearchIO
 
 
 @pytest.fixture()
@@ -67,7 +71,9 @@ def run_cblat(result_dir: Path, port: int, fa_data: Path):
     p.start()
     time.sleep(5)
 
-    for _, fa in enumerate(fa_data.glob("*.fa")):
+    fa_files = [fa_data] if fa_data.suffix == ".fa" else fa_data.glob("*.fa")
+
+    for _, fa in enumerate(fa_files):
         print(f"run cc {fa}")
         cc_res = result_dir / f"{fa.stem}_cc.psl"
         run_cmd(
@@ -76,6 +82,39 @@ def run_cblat(result_dir: Path, port: int, fa_data: Path):
 
     print("stop c server")
     run_cmd(f"{gfserver} stop localhost {cport}")
+
+
+def run_pxblat_async(result_dir: Path, port: int, fa_data: Path):
+    cport = port + 20
+    two_bit = Path("benchmark/data/chr20.2bit")
+    server_option = pxblat.ServerOption().withCanStop(True).withStepSize(5).build()
+
+    print("open python server")
+    server = pxblat.Server("localhost", cport, two_bit, server_option)
+    server.start()
+    server.wait_ready()
+    results = {}
+
+    client = Client(
+        host="localhost",
+        port=cport,
+        seq_dir=two_bit.parent.as_posix(),
+        min_score=20,
+        min_identity=90,
+    )
+
+    fa_files = (
+        [fa_data] if fa_data.suffix == ".fa" else [fa for fa in fa_data.glob("*.fa")]
+    )
+
+    ret = client.query(fa_files)
+
+    for r, fa in zip(ret, fa_files):
+        print(f"run px {fa}")
+        results[fa.stem] = r
+
+    server.stop()
+    return results
 
 
 def run_pxblat(result_dir: Path, port: int, fa_data: Path):
@@ -89,7 +128,8 @@ def run_pxblat(result_dir: Path, port: int, fa_data: Path):
     server.wait_ready()
     results = {}
 
-    for _, fa in enumerate(fa_data.glob("*.fa")):
+    fa_files = [fa_data] if fa_data.suffix == ".fa" else fa_data.glob("*.fa")
+    for _, fa in enumerate(fa_files):
         client_option = (
             pxblat.ClientOption()
             .withMinScore(20)
@@ -101,7 +141,7 @@ def run_pxblat(result_dir: Path, port: int, fa_data: Path):
             .build()
         )
 
-        client = pxblat.Client(client_option)
+        client = pxblat.ClientThread(client_option)
         client.start()
         ret = client.get()
         results[fa.stem] = ret
@@ -143,18 +183,25 @@ def get_overlap(hsps1, hsps2):
     return set1 - set2, set2 - set1, set1 & set2
 
 
-def _cpsl(file1, file2, isprint=True):
+def _cpsl(file1, file2, isprint=False):
     cc_psl = file1
     cp_psl = file2
 
-    from Bio import SearchIO
-
-    cc_res = SearchIO.read(cc_psl, "blat-psl")
+    cc_res = None
+    try:
+        cc_res = SearchIO.read(cc_psl, "blat-psl")
+    except ValueError as e:
+        if "No query results" in str(e):
+            print(f"No result found in {cc_psl}")
+            cc_res = None
 
     if isinstance(cp_psl, (Path, str)):
         cp_res = SearchIO.read(cp_psl, "blat-psl")
     else:
         cp_res = cp_psl
+
+    if cc_res is None and cp_res is None:
+        return
 
     cc_hsps = cc_res.hsps
     cp_hsps = cp_res.hsps
@@ -174,9 +221,9 @@ def _cpsl(file1, file2, isprint=True):
     return get_overlap(cc_hsps, cp_hsps)
 
 
-def create_result(result_dir, port, fas):
-    run_cblat(result_dir, port, fas)
-    results = run_pxblat(result_dir, port, fas)
+def create_result(result_dir, port, fa_data):
+    run_cblat(result_dir, port, fa_data)
+    results = run_pxblat_async(result_dir, port, fa_data)
     return results
 
 
@@ -187,35 +234,91 @@ def time_creat_result(result_dir, port, fas):
     print(f"run_cblat time: {cend - cstart:.4f}")
 
     pstart = time.perf_counter()
-    results = run_pxblat(result_dir, port, fas)
+    results = run_pxblat_async(result_dir, port, fas)
     pend = time.perf_counter()
     print(f"run_pxblat time: {pend - pstart:.4f}")
 
     speedup = (cend - cstart) / (pend - pstart)
-    print(f"speed up: {speedup:.4f}x")
+    print(f"\nspeed up: {speedup:.4f}x\n")
 
     return results
 
 
-def test_result(tmpdir, port, fas, time=False):
-    if not time:
-        pxblat_results = create_result(tmpdir, port, fas)
-    else:
-        pxblat_results = time_creat_result(tmpdir, port, fas)
+def test_result_c_with_py(tmpdir, port, fas, *, time=True, compare=True):
+    pxblat_results = (
+        create_result(tmpdir, port, fas)
+        if not time
+        else time_creat_result(tmpdir, port, fas)
+    )
 
-    for fa in fas.glob("*fa"):
+    if compare:
+        file_num = 0
+        for fa in fas.glob("*fa"):
+            file_num += 1
+            cc_res = tmpdir / f"{fa.stem}_cc.psl"
+            pp_res = pxblat_results[fa.stem]
+
+            _ret = _cpsl(cc_res, pp_res)
+            if _ret is None:
+                print(f"No result found {fa.stem}")
+                continue
+
+            a, b, _ = _ret
+            assert len(a) == 0
+            assert len(b) == 0
+        print(f"compare {file_num} files for BLAT and PxBLAT")
+
+
+@pytest.mark.skip
+@pytest.mark.parametrize(
+    "failing_fas",
+    [
+        Path("tests/data/fas/chr20_11648866_11650925.fa"),
+        Path("benchmark/fas/chr20_30806959_30809757.fa"),
+    ],
+)
+def test_failing_case(tmpdir, port, failing_fas, time=False):
+    pxblat_results = (
+        create_result(tmpdir, port, failing_fas)
+        if not time
+        else time_creat_result(tmpdir, port, failing_fas)
+    )
+
+    file_num = 0
+    for fa in [failing_fas]:
+        file_num += 1
         cc_res = tmpdir / f"{fa.stem}_cc.psl"
         pp_res = pxblat_results[fa.stem]
-        a, b, _ = _cpsl(cc_res, pp_res, False)
+
+        _ret = _cpsl(cc_res, pp_res)
+        if _ret is None:
+            return
+        a, b, _ = _ret
         assert len(a) == 0
         assert len(b) == 0
+
+    print(f"test {file_num} files")
 
 
 if __name__ == "__main__":
     from pathlib import Path
+    import sys
+
+    input_fa = Path(sys.argv[1])
+
+    if not input_fa.exists() and not input_fa.is_dir():
+        raise SystemExit(f"input file {input_fa} does not exist")
 
     tmpdir = Path("tmp")
     tmpdir.mkdir(exist_ok=True)
     port = 65000
     fa = Path("tests/data/fas/")
-    test_result(tmpdir, port, fa, True)
+    fa1 = Path("benchmark/fas/1")
+    fa2 = Path("benchmark/fas/2")
+    fa3 = Path("benchmark/fas/3")
+    fa4 = Path("benchmark/fas/4")
+    fa5 = Path("benchmark/fas/5")
+    fa6 = Path("benchmark/fas/6")
+    test_fas = Path("benchmark/test_fas")
+
+    test_result_c_with_py(tmpdir, port, input_fa, time=True, compare=True)